(refactor) move to utils
