+++ lib/gherlint/GherlintConfig.js
@@ -4,12 +4,12 @@ const glob = require("glob");
 const { isEmpty, merge, difference, keys } = require("lodash");
 
 const log = require("../logging/logger");
-const { isDir, cwd, searchFeatureFiles } = require("../Path");
+const { isDir, cwd, searchFeatureFiles } = require("../../utils/Path");
 const {
     configFilePattern,
     gherlintrc: defaultGherlintrc,
 } = require("./config");
-const Fs = require("../Fs");
+const Fs = require("../../utils/Fs");
 const ruleOptions = require("../rules/_ruleOptions");
 
 module.exports = class GherlintConfig {
